rearrange and clean up project
