+++ .gitignore
@@ -1,4 +1,6 @@
-node_modules/*
+node_modules
+dist
+
 accounts.json
 RRL_account.json
 RRL.json
