+++ RRL_releases.json
@@ -1,1 +1,12 @@
-{"6168":{"novelID":12147,"title":"Lorem Ipsum Ch.2 - Dolor","time":"2019-05-03T11:00:00.0000000"},"6169":{"novelID":12147,"title":"Lorem Ipsum Ch.3 - Sit Amet","time":"2019-05-10T11:00:00.0000000"}}
+{
+  "6168": {
+    "novelID": 12147,
+    "title": "Lorem Ipsum Ch.2 - Dolor",
+    "time": "2019-05-03T11:00:00.0000000"
+  },
+  "6169": {
+    "novelID": 12147,
+    "title": "Lorem Ipsum Ch.3 - Sit Amet",
+    "time": "2019-05-10T11:00:00.0000000"
+  }
+}
