--- accounts.gitignore
@@ -1,2 +1,0 @@
-# Don't upload the passwords to gitHub...
-accounts.json
