+++ index.html
@@ -1,39 +1,36 @@
-<!DOCTYPE html> 
+<!DOCTYPE html>
 <html>
-    <head>
-        <meta charset="utf-8"/>
-        <title>Chapter Coordinator</title>
-        <link rel="stylesheet" href="styles.css"/>
-    </head>
-    <body>
-        Chapter Coordinator Settings:
-        <table>
-            <tr>
-                <td>
-                    <div id="RRL_ScheduledReleases_container"></div>
-                    <div id="RRL_LocalChapters_container"></div>
-                    <div id="RRL_account_container"></div>
-                </td>
-            </tr>
-            <tr>
-                <td>
-                    <!--<div id="RRL_account"></div>   -->
-                    <div id="RRL_UpdateSettings_container"></div>         
-                    <div id="test_buttons_container"></div>         
-                </td>
-            </tr>
-        </table>
-        
-        
-    </body>
-    <script type="text/javascript" src="./renderer.js"></script>
-    <script type="module">
-        //function $(id){return document.getElementById(id);} 
-        require('./renderer.js');
-        window.RRL = require('./website.js');
-        window.form = require('./form_instances.js');
-        window.RRL.loadUpdateSettings();
-        var test_variable = "yo";     
-        window.form.updateAll();  
-    </script>
+  <head>
+    <meta charset="utf-8" />
+    <title>Chapter Coordinator</title>
+    <link rel="stylesheet" href="styles.css" />
+  </head>
+  <body>
+    Chapter Coordinator Settings:
+    <table>
+      <tr>
+        <td>
+          <div id="RRL_ScheduledReleases_container"></div>
+          <div id="RRL_LocalChapters_container"></div>
+          <div id="RRL_account_container"></div>
+        </td>
+      </tr>
+      <tr>
+        <td>
+          <!--<div id="RRL_account"></div>   -->
+          <div id="RRL_UpdateSettings_container"></div>
+          <div id="test_buttons_container"></div>
+        </td>
+      </tr>
+    </table>
+  </body>
+  <script>
+    //function $(id){return document.getElementById(id);}
+    require("./dist/renderer.js");
+    window.RRL = require("./dist/website.js");
+    window.form = require("./dist/form_instances.js");
+    window.RRL.loadUpdateSettings();
+    var test_variable = "yo";
+    window.form.updateAll();
+  </script>
 </html>
