--- modules.gitignore
@@ -1,2 +1,0 @@
-# Don't upload the huge node_modules folder
-node_modules/*
