+++ tsconfig.json
@@ -3,12 +3,19 @@
     "target": "es6",
     "module": "commonjs",
     "moduleResolution": "node",
+    "outDir": "dist",
+    "baseUrl": ".",
+    "paths": {
+      "*": ["node_modules/*"]
+    },
     "sourceMap": true,
+
     "emitDecoratorMetadata": true,
     "experimentalDecorators": true,
     "removeComments": false,
     "noImplicitAny": false,
-    "suppressImplicitAnyIndexErrors": true 
+    "suppressImplicitAnyIndexErrors": true
   },
-  "exclude": [ "node_modules" ] 
+  "exclude": ["node_modules"],
+  "include": ["src/**/*"]
 }
