--- updateStuff.js
@@ -1,12 +1,0 @@
-require('library.js')
-
-
-//Main script 
-
-//Start window
-
-//Scheduled execution: 
-//Loads Google docs object, if defined in settings file, and runs it. 
-//Loads RRL object, if defined in settings file, and runs it. 
-//Loads Patreon object, if defined in settings file, and runs it.  
-
